second_try

+++ services/movie_session.py
@@ -30,18 +30,18 @@ def get_movie_session_by_id(movie_session_id: int) -> MovieSession:
 
 def update_movie_session(
         session_id: int,
-        show_time: str = None,
-        movie_id: int = None,
-        cinema_hall_id: int = None
+        show_time: str | None = None,
+        movie_id: int | None = None,
+        cinema_hall_id: int | None = None
 ) -> MovieSession:
     try:
         movie_session = MovieSession.objects.get(id=session_id)
 
         if show_time:
             movie_session.show_time = show_time
-        if movie_id is not None:
+        if movie_id:
             movie_session.movie_id = movie_id
-        if cinema_hall_id is not None:
+        if cinema_hall_id:
             movie_session.cinema_hall_id = cinema_hall_id
 
         movie_session.save()
@@ -52,10 +52,10 @@ def update_movie_session(
 
 
 def delete_movie_session_by_id(movie_session_id: int) -> None:
-    try:
-        movie_session = MovieSession.objects.get(id=movie_session_id)
+    movie_session = MovieSession.objects.filter(id=movie_session_id)
+    if movie_session.exists():
         movie_session.delete()
-    except MovieSession.DoesNotExist:
+    else:
         raise ValueError(
             f"MovieSession with id {movie_session_id} does not exist."
         )
